Add note on keyId in the README
As a developer with no management rights to AWS IAM, I couldn't figure out whether I had to enter the keyId or not, so delayed some releases/commits until we had the keyID in the properties.

Turns out that we were using [AWS KMS key policies](http://docs.aws.amazon.com/kms/latest/developerguide/key-policies.html), and since the code was only decrypting, we didn't have to specify the keyId at all.

This pull request adds a sentence detailing that, and also the list of available/used properties. Maybe that could go under another section, so just let me know if the text needs to be updated, or moved elsewhere. Happy to update this pull request with further commits.